Allow non-normalized forms as input to NUDUPL
fmpz_gcdinv() requires that f is less than g, which correspond to f->b
and f->a coefficients of the input quadratic form, however this may not
be the case if the form is not normalized. Therefore NUDUPL did not work
correctly for such inputs.

Fix by reducing f modulo g in case f >= g.